rename the packages and modules to licenses
