+++ data/licenses/src/commonMain/kotlin/app/tivi/data/licenses/OpenSourceDataComponent.kt
@@ -1,0 +1,6 @@
+// Copyright 2017, Google LLC, Christopher Banes and the Tivi project contributors
+// SPDX-License-Identifier: Apache-2.0
+
+package app.tivi.data.licenses
+
+expect interface OpenSourceDataComponent
