+++ data/licenses/src/commonMain/kotlin/app/tivi/data/licenses/store/OpenSourceStore.kt
@@ -1,0 +1,13 @@
+// Copyright 2022, Google LLC, Christopher Banes and the Tivi project contributors
+// SPDX-License-Identifier: Apache-2.0
+
+package app.tivi.data.licenses.store
+
+import app.tivi.data.licenses.OpenSourceState
+
+interface OpenSourceStore {
+    fun fetch(): OpenSourceState
+    fun save(state: OpenSourceState)
+    fun clear()
+    fun isAvailable(): Boolean = true
+}
