--- data/opensource/src/commonMain/kotlin/app/tivi/data/opensource/OpenSourceDataComponent.kt
@@ -1,6 +1,0 @@
-// Copyright 2017, Google LLC, Christopher Banes and the Tivi project contributors
-// SPDX-License-Identifier: Apache-2.0
-
-package app.tivi.data.opensource
-
-expect interface OpenSourceDataComponent
