--- data/opensource/src/commonMain/kotlin/app/tivi/data/opensource/store/OpenSourceStore.kt
@@ -1,13 +1,0 @@
-// Copyright 2022, Google LLC, Christopher Banes and the Tivi project contributors
-// SPDX-License-Identifier: Apache-2.0
-
-package app.tivi.data.opensource.store
-
-import app.tivi.data.opensource.OpenSourceState
-
-interface OpenSourceStore {
-    fun fetch(): OpenSourceState
-    fun save(state: OpenSourceState)
-    fun clear()
-    fun isAvailable(): Boolean = true
-}
