+++ settings.gradle.kts
@@ -96,7 +96,7 @@ include(
     ":data:recommendedshows",
     ":data:relatedshows",
     ":data:search",
-    ":data:opensource",
+    ":data:licenses",
     ":data:shows",
     ":data:showimages",
     ":data:traktauth",
@@ -116,7 +116,7 @@ include(
     ":ui:episode:details",
     ":ui:episode:track",
     ":ui:trending",
-    ":ui:opensource",
+    ":ui:licenses",
     ":ui:popular",
     ":ui:recommended",
     ":ui:search",
